refactors handlers into different files
