+++ docs/usage/2-route-handlers/1_http_route_handlers.md
@@ -1,9 +1,4 @@
-# Route Handlers
-
-Route handlers are the core of Starlite. They are constructed by decorating a function or method with one of the handler
-decorators exported from Starlite.
-
-## HTTP Route Handlers
+# HTTP Route Handlers
 
 The base decorator is called `route`:
 
@@ -71,7 +66,7 @@ def my_endpoint() -> None:
     ...
 ```
 
-This is particularly useful when you want to have optional [path parameters](3-parameters.md#path-parameters):
+This is particularly useful when you want to have optional [path parameters](../3-parameters.md#path-parameters):
 
 ```python
 from starlite import HttpMethod, route
@@ -93,15 +88,15 @@ a list of members, e.g. `HttpMethod.GET` or `[HttpMethod.PATCH, HttpMethod.PUT]`
 Additionally, you can pass the following optional kwargs:
 
 - `status_code`: the status code for a success response. If not
-  specified, [a default value will be used](5-responses.md#status-codes).
+  specified, [a default value will be used](../5-responses.md#status-codes).
 - `media_type`: A string or a member of the enum `starlite.enums.MediaType`, which specifies the MIME Media Type for the
-  response. Defaults to `MediaType.JSON`. See [media-type](5-responses.md#media-type).
+  response. Defaults to `MediaType.JSON`. See [media-type](../5-responses.md#media-type).
 - `response_class`: A custom response class to be used as the app default.
-  See [using-custom-responses](5-responses.md#using-custom-responses).
+  See [using-custom-responses](../5-responses.md#using-custom-responses).
 - `response_headers`: A dictionary of `ResponseHeader` instances.
-  See [response-headers](5-responses.md#response-headers).
-- `dependencies`: A dictionary mapping dependency providers. See [dependency-injection](6-dependency-injection.md).
-- `opt`: String keyed dictionary of arbitrary value that can be used by [guards](9-guards.md).
+  See [response-headers](../5-responses.md#response-headers).
+- `dependencies`: A dictionary mapping dependency providers. See [dependency-injection](../6-dependency-injection.md).
+- `opt`: String keyed dictionary of arbitrary value that can be used by [guards](../9-guards.md).
 - `before_request`: A sync or async function to execute before a `Request` is passed to the route handler. If this
   function returns a value, the request will not reach the route handler, and instead this value will be used.
 - `after_request`: A sync or async function to execute before the `Response` is returned. This function receives the
@@ -111,7 +106,7 @@ Additionally, you can pass the following optional kwargs:
   the response is executed. Note - if you return a value from a `before_request` hook, background tasks passed to the
   handler will not be executed.
 
-And the following kwargs, which affect [OpenAPI schema generation](12-openapi.md#route-handler-configuration)
+And the following kwargs, which affect [OpenAPI schema generation](../12-openapi.md#route-handler-configuration)
 
 - `include_in_schema`: A boolean flag dictating whether the given route handler will appear in the generated OpenAPI
   schema. Defaults to `True`.
@@ -189,81 +184,6 @@ should be distinguished by a unique `operationId` and optimally also have a `sum
 As such, using the `route` decorator is discouraged. Instead, the preferred pattern is to share code using secondary
 class methods or by abstracting code to reusable functions.
 
-## Websocket Route Handlers
-
-!!! info
-    This feature is available from v0.2.0 onwards
-
-Alongside the HTTP Route handlers discussed above, Starlite also support Websockets via the `websocket` decorator:
-
-```python
-from starlite import WebSocket, websocket
-
-
-@websocket(path="/socket")
-async def my_websocket_handler(socket: WebSocket) -> None:
-    await socket.accept()
-    await socket.send_json({...})
-    await socket.close()
-```
-
-The `websocket` decorator is also an aliased class, in this case - of the `WebsocketRouteHandler`. Thus. you can write
-the above like so:
-
-```python
-from starlite import WebSocket, WebsocketRouteHandler
-
-
-@WebsocketRouteHandler(path="/socket")
-async def my_websocket_handler(socket: WebSocket) -> None:
-    await socket.accept()
-    await socket.send_json({...})
-    await socket.close()
-```
-
-In difference to HTTP routes handlers, websocket handlers have the following requirements:
-
-1. they **must** declare a `socket` kwarg. If this is missing an exception will be raised.
-2. they **must** have a return annotation of `None`. Any other annotation, or lack thereof, will raise an exception.
-
-Additionally, they should be async because the socket interface is async - but this is not enforced. You will not be
-able to do anything meaningful without this and python will raise errors as required.
-
-In all other regards websocket handlers function exactly like other route handlers.
-
-!!! note
-    OpenAPI currently does not support websockets. As a result not schema will be generated for websocket route
-    handlers, and you cannot configure any schema related parameters for these.
-
-## ASGI Route Handlers
-
-!!! info
-    This feature is available from v0.7.0 onwards
-
-You can write your own ASGI apps using the `asgi` route handler decorator:
-
-```python
-from starlette.types import Scope, Receive, Send
-from starlette.status import HTTP_200_OK, HTTP_400_BAD_REQUEST
-from starlite import Response, asgi
-
-
-@asgi(path="/my-asgi-app")
-async def my_asgi_app(scope: Scope, receive: Receive, send: Send) -> None:
-    if scope["type"] == "http":
-        if scope["method"] == "GET":
-            response = Response({"hello": "world"}, status_code=HTTP_200_OK)
-            await response(scope=scope, receive=receive, send=send)
-        return
-    response = Response(
-        {"detail": "unsupported request"}, status_code=HTTP_400_BAD_REQUEST
-    )
-    await response(scope=scope, receive=receive, send=send)
-```
-
-!!! note
-    ASGI apps are currently not handled in OpenAPI generation - although this might change in the future.
-
 ## Handler Function Kwargs
 
 Route handler functions or methods access various data by declaring these as annotated function kwargs. The annotated
@@ -271,9 +191,9 @@ kwargs are inspected by Starlite and then injected into the request handler.
 
 The following sources can be accessed using annotated function kwargs:
 
-1. [path, query, header and cookie parameters](3-parameters.md)
-2. [the request body](4-request-body.md)
-3. [dependencies](6-dependency-injection.md)
+1. [path, query, header and cookie parameters](../3-parameters.md)
+2. [the request body](../4-request-body.md)
+3. [dependencies](../6-dependency-injection.md)
 
 Additionally, you can specify the following special kwargs:
