+++ starlite/__init__.py
@@ -1,4 +1,3 @@
-# flake8: noqa
 from starlite.datastructures import File, Redirect, State, Stream, Template
 
 from .app import Starlite
@@ -50,25 +49,20 @@ from .testing import TestClient, create_test_client, create_test_request
 from .types import MiddlewareProtocol, Partial, ResponseHeader
 
 __all__ = [
-    "AbstractAuthenticationMiddleware",
-    "asgi",
     "ASGIRouteHandler",
+    "AbstractAuthenticationMiddleware",
     "AuthenticationResult",
     "BaseRoute",
     "BaseRouteHandler",
     "Body",
-    "Controller",
     "CORSConfig",
-    "create_test_client",
-    "create_test_request",
-    "delete",
+    "Controller",
     "DTOFactory",
     "File",
-    "get",
     "HTTPException",
-    "HttpMethod",
     "HTTPRoute",
     "HTTPRouteHandler",
+    "HttpMethod",
     "ImproperlyConfiguredException",
     "InternalServerException",
     "LoggingConfig",
@@ -82,31 +76,36 @@ __all__ = [
     "OpenAPIMediaType",
     "Parameter",
     "Partial",
-    "patch",
     "PermissionDeniedException",
     "PluginProtocol",
-    "post",
     "Provide",
-    "put",
     "Redirect",
     "Request",
     "RequestEncodingType",
     "Response",
     "ResponseHeader",
-    "route",
     "Router",
     "ScopeType",
     "ServiceUnavailableException",
-    "Starlite",
     "StarLiteException",
+    "Starlite",
     "State",
     "StaticFilesConfig",
     "Stream",
-    "TestClient",
     "Template",
     "TemplateConfig",
+    "TestClient",
     "WebSocket",
-    "websocket",
     "WebSocketRoute",
     "WebsocketRouteHandler",
+    "asgi",
+    "create_test_client",
+    "create_test_request",
+    "delete",
+    "get",
+    "patch",
+    "post",
+    "put",
+    "route",
+    "websocket",
 ]
